backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
httpx
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
alembic
boto3
Pillow
pillow-heif
exifread
imagehash
redis
python-multipart
slowapi
apscheduler
pgvector
